requirements.txt
streamlit
langchain-google-genai
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv